+++ src/components/Reuseable/index.js
@@ -1,3 +1,3 @@
-export ArcPieChart from './ArcPieChart/ArcPieChart';
+export HalfDonutChart from './HalfDonutChart/HalfDonutChart';
 export TitleNav from './TitleNav/TitleNav';
 export ListBarChart from './ListBarChart/ListBarChart';
